Using iframes to host my Github pages
Ironically, because React Router doesn't actually
direct the client to "real" urls, I can't route to
projects which are under pgrad.github.io directly.
pgrad.github.io/CS114_Final with React Router is not
the same as the actual static page. This means that
if I want to show the content in these pages within
the React app, I need to use an iframe to host the
external page.