Make the website a PWA
To create a PWA, I have to add a
manifest.json which defines what route
the app starts at, what icons to show,
and other info.

I also have to add a service worker
which will attempt to load the page,
but if there's no internet access it
loads an offline page. Given that this
website is a single page app, I just
link the offline page back to the
index.html. I'll have to actually
implement some logic for offline
behavior.